Cleaned up data files and linked them locally, rather than inside components
