--- src/data/dirsEdgerunnerFTP.js
@@ -1,9 +1,0 @@
-import { Directory } from "../functions/classes"
-
-export const edgeDirs = {
-    rootDir: new Directory("root"),
-    homeDir: new Directory("home"),
-    archiveDir: new Directory("archive"),
-    softwareDir: new Directory("software"),
-    targetsDir: new Directory("targets")
-}
