--- src/data/reso.js
@@ -1,173 +1,0 @@
-import { File } from "../functions/classes";
-import { net } from "./network";
-
-let username = "";
-const getUsername = () => {
-  if (!window.localStorage.getItem("username")) {
-    return "";
-  } else {
-    const userID = window.localStorage.getItem("username").split("@");
-    return (username = userID[0]);
-  }
-};
-
-getUsername();
-
-export const reso = {
-  multi: {
-    permissions: new File(
-      "permission denied",
-      "text",
-      "root",
-      "root",
-      "you do not have permission to access this directory",
-      "n/a"
-    ),
-  },
-  root: {
-    userLog0: new File(
-      "connections-logs_0",
-      "text",
-      "storage",
-      "tiNeptune",
-      "tiNeptune",
-      "New user " +
-        username +
-        " spawned shell from " +
-        net.zombie._ipAddress +
-        ". Latency geolocation cache connection location: Pallo Alto, California",
-      "n/a"
-    ),
-    userLog1: new File(
-      "connections-logs_1",
-      "text",
-      "storage",
-      "tiNeptune",
-      "tiNeptune",
-      username + "@" + net.zombie._ipAddress + ": password authenticated successfully"
-    ),
-  },
-  bbs_mask: {
-    a91EOC5NMg5xE: new File(
-      "91E0C5NMg5xE",
-      "encrypted archive",
-      "root storage network",
-      "root",
-      "root",
-      "JKJSxHQ~3kLe[B*6d9AslIly2B4UZbhiRd3~Pr^HzdSeoFprA1~0DG;9K1'HQUUQ",
-      "n/a"
-    ),
-    bHjTRSlk7aw4b: new File(
-      "HjTRSlk7aw4b",
-      "encrypted archive",
-      "root storage network",
-      "root",
-      "root",
-      "zQungkqNf*j?-TQ0~?q[rHEXWDph~vk-UHef8[g1Pm|~ucVJcDloQ:E::Wht;ej-",
-      "n/a"
-    ),
-    cM9KFZKUlring: new File(
-      "M9KFZKUlring",
-      "encrypted archive",
-      "root storage network",
-      "root",
-      "root",
-      "5GuLEnMe|9[x3%riy]QQyx~p_T,Wh1:AzYzcjg[!*'_4!~FV?#0Cy?;,o'+4p+[",
-      "n/a"
-    ),
-    esaoXGjvToAlh: new File(
-      "saoXGjvToAlh",
-      "encrypted archive",
-      "root network i/o-bus gpg-sha",
-      "root tiNeptune",
-      "root",
-      "DcfkzA2v~0OU7:N**VFV+|O8ZqG~[RK1Dl!s8IXuHZ:_tcOdkktA4GmE,h[O#Z_",
-      "pa tèlman parese tan sa a, gen aksè konfigirasyon"
-    ),
-  },
-  p7FzTKa13Ohg: {
-    ruleset: new File(
-      "ruleset",
-      "text",
-      "storage network cron",
-      "root tiNeptune",
-      "root",
-      "tcp/ip i/o = in | udp i/o = i/o | bbs mask protocol = sha2048{key = root/var/sha.d/lA&yhTW-.hash}",
-      "n/a"
-    ),
-    ufwConf: new File(
-      "ufw.conf.d",
-      "text",
-      "root storage network cron",
-      "root tiNeptune",
-      "root",
-      "tcp/udp-ip: port 5133 = open | tcp/udp-ip: port 1-5132 = closed && tcp/udp-ip: port 5134-* = closed",
-      "n/a"
-    ),
-  },
-  whitelist: {
-    jeonKiri: new File(
-      "jeon-kiri",
-      "text",
-      "storage network",
-      "root tiNeptune",
-      "tiNeptune",
-      "static ip: 251.45.27.176",
-      "n/a"
-    ),
-    leon: new File(
-      "leon",
-      "text",
-      "storage network",
-      "root tiNeptune",
-      "tiNepture",
-      "static ip: 56.35.227.238",
-      "n/a"
-    ),
-    mamanBriggite: new File(
-      "maman-briggite",
-      "text",
-      "storage network",
-      "root tiNeptune",
-      "tiNeptune",
-      "static ip: 68.225.7.23",
-      "n/a"
-    ),
-    moseley: new File(
-      "moseley",
-      "text",
-      "storage network",
-      "root",
-      "root",
-      "static ip: 106.14.156.224",
-      "casefile:cYt]09*n!,H+ operation: golden jackal"
-    ),
-    mrHands: new File(
-      "mr-hands",
-      "text",
-      "storage network",
-      "root tiNeptune",
-      "tiNeptune",
-      "static ip: 226.126.32.203",
-      "n/a"
-    ),
-    placide: new File(
-      "placide",
-      "text",
-      "storage network",
-      "root tiNeptune",
-      "tiNeptune",
-      "static ip: 211.98.124.26",
-      "n/a"
-    ),
-    tiNeptune: new File(
-      "ti-neptune",
-      "text",
-      "storage network",
-      "root tiNeptune",
-      "tiNeptune",
-      "static ip: 36.155.4.191",
-      "n/a"
-    ),
-  },
-};
